Fixed inconsistency in the creation of :wiki-url from :web-home URL.
The :wiki-url built for the namespace-index-info does not insert a "/"
between the :web-home and the namespace file name (it assumes the
given :web-home has a trailing "/").
The :wiki-url built for the var-index-info does insert a "/" between
the :web-home and the var file name (it assumes the given :web-home
does not end in "/").

You can easily see it in the generated index.clj. Search for :wiki-url.
Either the ones in the :namespaces section have one "/" correctly and
the ones in the :vars section have a "//" in the middle of the URL, or
the former are missing the "/" and the latter are correct.

Inserting an extra "/" in the code for namespace-index-info fixed the
problem. If :web-home does not have a trailing "/", the resulting
:wiki-url for namespaces and vars are correct; if :web-home ends with
a "/" the resulting :wiki-url will contain two consecutive "//", but
still work correctly.